auth/raspberry/requirements.txt
paho-mqtt==1.6.1
pydantic==1.9.0
python-dotenv==0.20.0
setuptools==57.0.0
six==1.16.0
typing-extensions==4.1.1
wheel==0.36.2
